requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
redis==5.0.1
kafka-python==2.0.2
pydantic==2.9.0
python-dotenv==1.0.1
psycopg2-binary
